Separate ClosestNotMeConvexResultCallback
