+++ apps/openmw/mwphysics/trace.cpp
@@ -7,47 +7,10 @@
 
 #include "collisiontype.hpp"
 #include "actor.hpp"
+#include "closestnotmeconvexresultcallback.hpp"
 
 namespace MWPhysics
 {
-
-class ClosestNotMeConvexResultCallback : public btCollisionWorld::ClosestConvexResultCallback
-{
-public:
-    ClosestNotMeConvexResultCallback(const btCollisionObject *me, const btVector3 &motion, btScalar minCollisionDot)
-      : btCollisionWorld::ClosestConvexResultCallback(btVector3(0.0, 0.0, 0.0), btVector3(0.0, 0.0, 0.0)),
-        mMe(me), mMotion(motion), mMinCollisionDot(minCollisionDot)
-    {
-    }
-
-    virtual btScalar addSingleResult(btCollisionWorld::LocalConvexResult& convexResult,bool normalInWorldSpace)
-    {
-        if(convexResult.m_hitCollisionObject == mMe)
-            return btScalar( 1 );
-
-        btVector3 hitNormalWorld;
-        if(normalInWorldSpace)
-            hitNormalWorld = convexResult.m_hitNormalLocal;
-        else
-        {
-            ///need to transform normal into worldspace
-            hitNormalWorld = convexResult.m_hitCollisionObject->getWorldTransform().getBasis()*convexResult.m_hitNormalLocal;
-        }
-
-        // dot product of the motion vector against the collision contact normal
-        btScalar dotCollision = mMotion.dot(hitNormalWorld);
-        if(dotCollision <= mMinCollisionDot)
-            return btScalar(1);
-
-        return ClosestConvexResultCallback::addSingleResult(convexResult, normalInWorldSpace);
-    }
-
-protected:
-    const btCollisionObject *mMe;
-    const btVector3 mMotion;
-    const btScalar mMinCollisionDot;
-};
-
 
 void ActorTracer::doTrace(const btCollisionObject *actor, const osg::Vec3f& start, const osg::Vec3f& end, const btCollisionWorld* world)
 {
